requirements.txt
# Основные зависимости для psql-tools
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
peewee>=3.16.0
playhouse>=0.1.0
